Fix line endings
