+++ service-print/src/main/java/org/oskari/print/PNG.java
