+++ service-print/src/main/java/org/oskari/print/wmts/TileMatrixSetCache.java
@@ -42,7 +42,7 @@ public class TileMatrixSetCache {
             if (tileMatrixSets != null) {
                 for (TileMatrixSet tileMatrixSet : tileMatrixSets) {
                     String key = getKey(layer.getId(), tileMatrixSet.getId());
-                    LOG.debug("Adding {} to cache", key);
+                    LOG.debug("Adding", key, "to cache");
                     CACHE.put(key, tileMatrixSet);
                 }
             }
