+++ service-print/src/main/java/org/oskari/print/wmts/WMTSTileMatrixSetParser.java
@@ -10,7 +10,7 @@ import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 
-import org.oskari.util.DOMHelper;
+import org.oskari.print.util.DOMHelper;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.xml.sax.SAXException;
