+++ service-print/src/test/java/org/oskari/print/PrintFormatTest.java
@@ -5,7 +5,7 @@ import org.junit.Test;
 import org.oskari.print.PrintFormat;
 
 public class PrintFormatTest {
-    
+
     @Test
     public void nullContentTypeReturnsNull() {
         PrintFormat format = PrintFormat.getByContentType(null);
@@ -17,21 +17,21 @@ public class PrintFormatTest {
         PrintFormat format = PrintFormat.getByContentType("foo");
         Assert.assertNull(format);
     }
-    
+
     @Test
     public void knownFormatsReturnThemselves() {
         PrintFormat expected = PrintFormat.PDF;
         PrintFormat actual = PrintFormat.getByContentType(expected.contentType);
         Assert.assertEquals(expected, actual);
     }
-    
+
     @Test
     public void casingOfContentTypeDoesNotMatter() {
         PrintFormat expected = PrintFormat.PDF;
         String type = expected.contentType.toUpperCase();
         PrintFormat actual = PrintFormat.getByContentType(type);
         Assert.assertEquals(expected, actual);
-        
+
         type = type.toLowerCase();
         actual = PrintFormat.getByContentType(type);
         Assert.assertEquals(expected, actual);
